+++ custats/backend/functions.cu
@@ -4,31 +4,31 @@
 #include "kernels.h"
 #include "functions.h"
 
-void poisson_logpmf_hh2h(const int *k, const double *r, double *out, const int size)
+void poisson_logpmf_hh2h(const int *k, const float *r, float *out, const int size)
 {
   int *k_d;
-  double *r_d, *out_d;
+  float *r_d, *out_d;
   cuda_errchk(cudaMalloc(&k_d, size*sizeof(int)));
-  cuda_errchk(cudaMalloc(&r_d, size*sizeof(double)));
-  cuda_errchk(cudaMalloc(&out_d, size*sizeof(double)));
+  cuda_errchk(cudaMalloc(&r_d, size*sizeof(float)));
+  cuda_errchk(cudaMalloc(&out_d, size*sizeof(float)));
   cuda_errchk(cudaMemcpy(k_d, k, size*sizeof(int), cudaMemcpyHostToDevice));
-  cuda_errchk(cudaMemcpy(r_d, r, size*sizeof(double), cudaMemcpyHostToDevice));
+  cuda_errchk(cudaMemcpy(r_d, r, size*sizeof(float), cudaMemcpyHostToDevice));
 
   kernels::call_poisson_logpmf_kernel(k_d, r_d, out_d, size);
 
-  cuda_errchk(cudaMemcpy(out, out_d, size*sizeof(double), cudaMemcpyDeviceToHost));
+  cuda_errchk(cudaMemcpy(out, out_d, size*sizeof(float), cudaMemcpyDeviceToHost));
 }
 
-void poisson_logpmf_dh2d(const int *k, const double *r, double *out, const int size)
+void poisson_logpmf_dh2d(const int *k, const float *r, float *out, const int size)
 {
-  double *r_d;
-  cuda_errchk(cudaMalloc(&r_d, size*sizeof(double)));
-  cuda_errchk(cudaMemcpy(r_d, r, size*sizeof(double), cudaMemcpyHostToDevice));
+  float *r_d;
+  cuda_errchk(cudaMalloc(&r_d, size*sizeof(float)));
+  cuda_errchk(cudaMemcpy(r_d, r, size*sizeof(float), cudaMemcpyHostToDevice));
 
   kernels::call_poisson_logpmf_kernel(k, r_d, out, size);
 }
 
-void poisson_logpmf_hd2d(const int *k, const double *r, double *out, const int size)
+void poisson_logpmf_hd2d(const int *k, const float *r, float *out, const int size)
 {
   int *k_d;
   cuda_errchk(cudaMalloc(&k_d, size*sizeof(int)));
@@ -37,7 +37,7 @@ void poisson_logpmf_hd2d(const int *k, const double *r, double *out, const int s
   kernels::call_poisson_logpmf_kernel(k_d, r, out, size);
 }
 
-void poisson_logpmf_dd2d(const int *k, const double *r, double *out, const int size)
+void poisson_logpmf_dd2d(const int *k, const float *r, float *out, const int size)
 {
   kernels::call_poisson_logpmf_kernel(k, r, out, size);
 }
